cmd/compile/internal/syntax: improve error msg for unnamed param decl
Improve error message for unnamed param decl in identifier list.

When the named and unnamed parameter declarations are mixed together,
the parser should report an error because we expect an identifier list.

The parser assumes the symbol is a unnamed parameter and treats it as a type.
This is not always correct and parser has no ablity to determine this actually.

Reporting the named parameter declaration is incomplete is better.

fixes: golang/go#69506
Change-Id: I22969e84d1872fb70ef42389af44bd2e6f5a41ea