go/parser: improve error message for unnamed param decl in identifier list
When the named and unnamed parameter declarations are mixed together,
the parser assumes the symbol in unnamed parameter is a type,
and reports the error for the type.

This is not always correct and parser has no ablity to determine this actually.
Reporting the named parameter declaration is incomplete is better.

Updates: golang/go#69506

Change-Id: Ic3f29dc27067c881c2cc7954bc51a7653d6a41d2