crypto/internal/fips/sha3: import x/crypto/sha3
For now just internally, pending a dedicated proposal for the exposed
package API.

In this CL the code is copied verbatim, for ease of review. Only the
imports were replaced with the corresponding internal ones, and
crypto.RegisterHash calls were disabled.

DO NOT SUBMIT until CL 616555 is submitted, and this CL is synced, then
specify here what commit was imported.

Updates #65269
For #69536

Change-Id: Ia4735b50c99b9573a5c4889733c4a119930fe658