make.bash,make.bat: hint how to install the bootstrap Go
Whenever we upgrade the bootstrap version of Go, I spend at least 5
minutes trying to remember/find the command to download a new version
of Go. This CL prints that command as part of the error message.

I spent a little while trying to make the message more informative.
Ideally, it would look for "go1.*" in PATH, but there seems to be no
portable way to do this. Alternatively, it could look for
$HOME/sdk/go1.*, but then does it print a potentially long path to the
command, or just use this as a version hint and then look in PATH for
a short command? And if it doesn't find that, does it still print a
more generic hint? I decided to keep it all simple.

Change-Id: I0c4ed250a13e5d6866432dee5bd207b607d45559