syscall: unconditionally set FILE_FLAG_BACKUP_SEMANTICS in Open
`FILE_FLAG_BACKUP_SEMANTICS` is necessary for opening directories.
It is also useful for services doing backup tasks, but that use case
is not completely covered unless that flag is set unconditionally.

Note that setting `FILE_FLAG_BACKUP_SEMANTICS` is no-op (except for
the fact that it is required for opening directories) if the
`SE_BACKUP_NAME` and `SE_RESTORE_NAME` privileges are not held by the
process.

Fixes #23312.

Change-Id: I20a516047eadee3b37ce41f6fb8002456b3f3aa6