runtime: unify lock2, allow deeper sleep
The tri-state mutex implementation (unlocked, locked, sleeping) avoids
sleep/wake syscalls when contention is low or absent, but its
performance degrades when many threads are contending for a mutex to
execute a fast critical section.

A fast critical section means frequent unlock2 calls. Each of those
finds the mutex in the "sleeping" state and so wakes a sleeping thread,
even if many other threads are already awake and in the spin loop of
lock2 attempting to acquire the mutex for themselves. Many spinning
threads means wasting energy and CPU time that could be used by other
processes on the machine. Many threads all spinning on the same cache
line leads to performance collapse.

Merge the futex- and semaphore-based mutex implementations by creating a
semaphore abstraction for futex platforms. Then, add a bit to the mutex
state word that communicates whether one of the waiting threads is awake
and spinning. When threads in lock2 see the new "spinning" bit, they can
sleep immediately. In unlock2, the "spinning" bit means we can save a
syscall and not wake a sleeping thread.

This brings up the real possibility of starvation: waiting threads are
able to enter a deeper sleep than before, since one of their peers can
volunteer to be the sole "spinning" thread and thus cause unlock2 to
skip the futexwake syscall. It may now be easier for a small set of
threads to capture a mutex, with the other threads sleeping through an
unbounded number of unlock2 calls.

This change perpetuates the longstanding design choice of lock_sema.go,
for the threads that are asleep while waiting for a mutex form a LIFO
stack. For contended mutex values where the critical section is long
relative to the amount of time a thread spins before sleeping, it's long
been the case (on lock_sema.go platforms) that threads at the bottom of
that stack would sleep through an unbounded number of unlock2 calls.

However, without the new "spinbit" mechanism, a sufficiently short
critical section would lead to enough wakeups per second that none of
the threads would sleep for long at all, so it's not clear whether
lock_sema.go's LIFO behavior has shown up much in practice.

Thus, this deep sleep may be a new source of tail latency for programs
that operate close to the capacity of individual mutex values.

Add a new GOEXPERIMENT, "spinbitmutex", so it's easy to disable this new
approach. Enable it by default on supported platforms (the efficient
implementation requires atomic.Xchg8).

For #68578

    goos: linux
    goarch: amd64
    pkg: runtime
    cpu: 13th Gen Intel(R) Core(TM) i7-13700H
                                │      old       │                  new                  │
                                │     sec/op     │    sec/op      vs base                │
    MutexContention                 17.84n ±   0%   17.77n ±   0%   -0.36% (p=0.000 n=10)
    MutexContention-2               19.53n ±  23%   22.38n ±  13%        ~ (p=0.189 n=10)
    MutexContention-3               23.73n ±   2%   21.66n ±   7%   -8.70% (p=0.001 n=10)
    MutexContention-4               28.41n ±   3%   22.62n ±   7%  -20.35% (p=0.000 n=10)
    MutexContention-5               31.88n ±   2%   22.57n ±   5%  -29.23% (p=0.000 n=10)
    MutexContention-6               34.79n ±   3%   23.05n ±   5%  -33.73% (p=0.000 n=10)
    MutexContention-7               43.60n ±   4%   21.92n ±   2%  -49.72% (p=0.000 n=10)
    MutexContention-8               54.23n ±   1%   24.26n ±   7%  -55.26% (p=0.000 n=10)
    MutexContention-9               64.86n ±   2%   23.29n ±   5%  -64.10% (p=0.000 n=10)
    MutexContention-10              75.12n ±   5%   23.36n ±   5%  -68.91% (p=0.000 n=10)
    MutexContention-11              89.76n ±   4%   23.11n ±  11%  -74.26% (p=0.000 n=10)
    MutexContention-12              99.12n ±   4%   23.75n ±   4%  -76.04% (p=0.000 n=10)
    MutexContention-13             112.20n ±   3%   23.77n ±   4%  -78.81% (p=0.000 n=10)
    MutexContention-14             124.20n ±   3%   23.21n ±   6%  -81.32% (p=0.000 n=10)
    MutexContention-15             124.60n ±   7%   24.48n ±  11%  -80.35% (p=0.000 n=10)
    MutexContention-16             127.90n ±   7%   24.70n ±   5%  -80.69% (p=0.000 n=10)
    MutexContention-17             148.55n ±  10%   24.70n ±   4%  -83.38% (p=0.000 n=10)
    MutexContention-18             147.60n ±   3%   24.92n ±   4%  -83.12% (p=0.000 n=10)
    MutexContention-19             154.40n ±  11%   25.45n ±   8%  -83.52% (p=0.000 n=10)
    MutexContention-20             156.80n ±  12%   23.04n ±   2%  -85.30% (p=0.000 n=10)
    MutexHandoff/Solo-20            13.59n ±   1%   13.58n ±   3%        ~ (p=0.955 n=10)
    MutexHandoff/FastPingPong-20    138.2n ± 130%   135.2n ± 116%        ~ (p=0.796 n=10)
    MutexHandoff/SlowPingPong-20    1.783µ ±  14%   1.778µ ±   7%        ~ (p=0.912 n=10)
    geomean                         74.22n          29.49n         -60.27%

    goos: darwin
    goarch: arm64
    pkg: runtime
    cpu: Apple M1
                                │      old       │                 new                  │
                                │     sec/op     │    sec/op     vs base                │
    MutexContention               13.48n ±    4%   12.26n ±  5%   -9.05% (p=0.000 n=10)
    MutexContention-2             15.43n ±    2%   15.32n ±  2%        ~ (p=0.086 n=10)
    MutexContention-3             19.00n ±    2%   16.04n ±  2%  -15.56% (p=0.000 n=10)
    MutexContention-4             20.50n ±    1%   16.58n ±  0%  -19.08% (p=0.000 n=10)
    MutexContention-5             23.30n ±    1%   16.53n ±  1%  -29.06% (p=0.001 n=10)
    MutexContention-6             25.11n ±    0%   16.54n ±  2%  -34.14% (p=0.000 n=10)
    MutexContention-7             26.39n ±    1%   16.59n ±  2%  -37.14% (p=0.000 n=10)
    MutexContention-8             25.91n ±    0%   17.80n ± 38%  -31.30% (p=0.001 n=10)
    MutexHandoff/Solo-8           13.46n ±    5%   11.65n ±  6%  -13.38% (p=0.001 n=10)
    MutexHandoff/FastPingPong-8   66.54n ± 8148%   63.97n ±  3%   -3.87% (p=0.043 n=10)
    MutexHandoff/SlowPingPong-8   25.98µ ±   43%   26.95µ ± 17%        ~ (p=0.684 n=10)
    geomean                       42.19n           34.45n        -18.35%

Change-Id: I0058c75eadf282d08eea7fce0d426f0518039f7c