cmd/compile: allow adding intrinsic aliases before the intrinsic itself
When the existing code adds an alias for a intrinsic on a set of architectures,
it will not panic as long as the alias is successfully added on at least one
architecture, even if the intrinsic does not exist on other architectures.

After this CL, we can manage intrinsic aliases more precisely, and the code may
be more intuitive.

Change-Id: I09bf12e2c6fceda80f6bb6902a61b60d9eb0ff03