crypto/internal/fips: add EDDSA ACVP test coverage
This commit adds ACVP test coverage for EDDSA (Ed25519, and
HashEd25519/Ed25519ph) for the keyGen, keyVer, sigGen, and sigVer
capabilities.

Since we haven't yet moved the EDDSA code into the fips package this is
a **work in progress** and testing the pre-existing public EDDSA and
packages from crypto.

It is also using a fork of acvptool pending a BoringSSL CR being merged
to add EDDSA support to the tooling.

Change-Id: I5122d86180bd4d2f7d94570a6dc939808aa24fc4