spec: remove notion of core types
DO NOT SUBMIT.

This removes the notion of core types from the spec
in all instances except for type inference and unification.

Instead of referring to core types, each section that did
so before is reverted to approx. the pre-generics (1.17)
prose, and additional paragraphs cover the type parameter
case as needed.

The hope is that this makes it easier to read the spec.
When type parameters are involved, the extra prose is
local to the language feature in question and thus more
readily available. It also makes it possible to fine-tune
behavior for each such language feature.

Work in progress.

Experiment, requires proposal.

Change-Id: I2ca3c10443acbb4b97869b606510416823d60d8b