cmd/compile: fix wrong escape analysis for nested rangefunc
The closures for body of range functions within function f are named
like "f-range1", "f-range2", ..., "f-rangeN". If "f-range2" is contained
within "f-range1", the escape analysis will incorrectly see that
"f-range2" is not, since it expects the parent function name as part of
the closure name.

To fix this, containsClosure must be updated to handle the generated
names for range functions.

Fixes #70035

Change-Id: If7b138d1b0e2bef7f0c7d7df6e5eacd2ba7eae48